requirements.txt
graphql-core
requests